decorator置き換え
