+++ app/decorators/goal_decorator.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+module GoalDecorator
+end
