+++ app/helpers/application_helper.rb
@@ -19,10 +19,4 @@ module ApplicationHelper
     @selected_subgoal ||= selected_goal.subgoals.find_by_id(session[:subgoal_id]) if session[:subgoal_id]
   end
 
-  def subgoal_deadline(object)
-    select = {'１ヶ月後'=> 1, '２ヶ月後'=> 2, '３ヶ月後'=> 3, '４ヶ月後'=> 4, '５ヶ月後'=> 5, '半年後'=> 6, 'これ以上期限を延ばす場合は目標の期限を延ばしてください' => 0 }
-    # goalをもとに何ヶ月伸ばせるかがわかる
-    deadline = @subgoal.delivery_time_to_month(object)
-    select_form = select.select{|key, value| deadline.to_i >= value.to_i}
-  end
 end
