+++ app/models/application_record.rb
@@ -2,13 +2,11 @@ class ApplicationRecord < ActiveRecord::Base
   self.abstract_class = true
 
   def when_deadline(month)
-    byebug
     if self.persisted?
       # 締切日から❍ヶ月後
       self.deadline_on = months_later(month: month, time: deadline_on)
       # ここに余りがあれば足す処理追加
       if self.respond_to?(:division_remainder)
-        #self.deadline_on = self.deadline_on.since((self.division_remainder.to_i).days)
         self.deadline_on = days_later(day: division_remainder, time: deadline_on)
       end
       deadline_on
