+++ spec/decorators/goal_decorator_spec.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+require 'rails_helper'
+
+RSpec.describe GoalDecorator do
+  let(:goal) { Goal.new.extend GoalDecorator }
+  subject { goal }
+  it { should be_a Goal }
+end
